rename camerax to camera module with camerax as child package
