+++ camera/proguard-rules.pro
